analyzer/requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
psycopg2-binary==2.9.9
matplotlib==3.8.2
seaborn==0.13.2